+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `CommandError` and `ConsoleApplicationError.CommandError`
 
 ## 4.1.0 - 2022-10-24
 - Use `ConsoleStyle` `4.2`